fix(game): BG_G2ClientNeckAngles — reproduce C's unsuffixed-double literal arithmetic
Raven's thoracic/head angle expressions use UNSUFFIXED double literals
(0.4/0.1/0.6) — C promotes to f64, computes, and narrows once on store —
while the neck lines use float literals (0.2f/0.3f). Our port flattened
all of them to f32, a 1-ULP divergence in the cranium/thoracic bone-angle
overrides fed to the engine skeleton every frame.

Found by the lockstep referee's argument-level hunt (2026-07-14): the
G_G2_ANGLEOVERRIDE float args (never hashed by the syscall digest)
diverged by exactly 1 ULP at t=10250 — the session's true first
divergence, 80+ seconds before the first digest-visible event. Post-fix,
the first 22,646 bone-angle calls are bit-identical (was: 20).

Known-remaining (same hunt, next target): a look-target branch divergence
at a respawn frame — the renderInfo look chain (lookTarget/eyePoint/
lookingDebounceTime, all module-private) needs its own oracle comparison;
the frame-1806 halt signature persists until then.

Gates: six referee scenarios byte-identical, workspace tests green, fmt
clean. [push held]

+++ crates/mp/game/src/bg_pmove.rs
@@ -8822,30 +8822,36 @@ pub fn BG_G2ClientNeckAngles(
     }
 
     //split it up between the neck and cranium
+    // Raven's `0.4`/`0.1`/`0.6` are UNSUFFIXED double literals (the `0.5f` and
+    // neck `0.2f`/`0.3f` are floats): the thoracic/head expressions run in
+    // f64 and narrow once on store. f32-flattening here was a 1-ULP cranium
+    // divergence (lockstep t=10250 find, 2026-07-14).
     if thoracicAngles[PITCH] != 0.0 {
         //already been set above, blend them
-        thoracicAngles[PITCH] = (thoracicAngles[PITCH] + (lA[PITCH] * 0.4)) * 0.5;
+        thoracicAngles[PITCH] =
+            ((thoracicAngles[PITCH] as f64 + (lA[PITCH] as f64 * 0.4)) * 0.5) as f32;
     } else {
-        thoracicAngles[PITCH] = lA[PITCH] * 0.4;
+        thoracicAngles[PITCH] = (lA[PITCH] as f64 * 0.4) as f32;
     }
     if thoracicAngles[YAW] != 0.0 {
-        thoracicAngles[YAW] = (thoracicAngles[YAW] + (lA[YAW] * 0.1)) * 0.5;
+        thoracicAngles[YAW] = ((thoracicAngles[YAW] as f64 + (lA[YAW] as f64 * 0.1)) * 0.5) as f32;
     } else {
-        thoracicAngles[YAW] = lA[YAW] * 0.1;
+        thoracicAngles[YAW] = (lA[YAW] as f64 * 0.1) as f32;
     }
     if thoracicAngles[ROLL] != 0.0 {
-        thoracicAngles[ROLL] = (thoracicAngles[ROLL] + (lA[ROLL] * 0.1)) * 0.5;
+        thoracicAngles[ROLL] =
+            ((thoracicAngles[ROLL] as f64 + (lA[ROLL] as f64 * 0.1)) * 0.5) as f32;
     } else {
-        thoracicAngles[ROLL] = lA[ROLL] * 0.1;
+        thoracicAngles[ROLL] = (lA[ROLL] as f64 * 0.1) as f32;
     }
 
     neckAngles[PITCH] = lA[PITCH] * 0.2;
     neckAngles[YAW] = lA[YAW] * 0.3;
     neckAngles[ROLL] = lA[ROLL] * 0.3;
 
-    headAngles[PITCH] = lA[PITCH] * 0.4;
-    headAngles[YAW] = lA[YAW] * 0.6;
-    headAngles[ROLL] = lA[ROLL] * 0.6;
+    headAngles[PITCH] = (lA[PITCH] as f64 * 0.4) as f32;
+    headAngles[YAW] = (lA[YAW] as f64 * 0.6) as f32;
+    headAngles[ROLL] = (lA[ROLL] as f64 * 0.6) as f32;
 
     unsafe {
         strap_G2API_SetBoneAngles(
